Move all methods out of values
